Fixed jumping bug with click to drag, started working on click to select
The findHexTile method is messy since I am trying 3 different methods
and don't want to lose the code for the previous methods. It doesn't
quite work yet... It gets somewhat close but the math is tricky. The
code for a tile to get selected outside of that method does work though.